+++ empire/server/modules/powershell/situational_awareness/network/get_sql_server_login_default_pw.py
@@ -54,4 +54,5 @@ class Module:
             script_end=script_end,
             obfuscate=obfuscate,
             obfuscation_command=obfuscation_command,
+            script_already_obfuscated=obfuscate,
         )